moving actual node-t scripts to lib/ folder
