+++ lib/widgets.js
